fix: api route file struc
